requirements.txt
jupyter==1.0.0
matplotlib==3.1.0
numpy==1.17.3
pandas==0.25.1
seaborn==0.9.0
Scrapy==1.8.0
mysql-connector-python==8.0.19